requirement.txt
apify-client==2.4.1
arxiv==2.4.1
beautifulsoup4==4.12.3
faiss-cpu==1.13.2
feedparser==6.0.11
langchain==1.2.10
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.17
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langserve==0.3.3
numpy==1.26.4
pandas==2.2.2
pypdf==6.8.0
PyPDF2==3.0.1
python-dotenv==1.2.2
scikit-learn==1.5.1
sentence-transformers==2.7.0
streamlit>=1.32.0
tiktoken==0.7.0
transformers==4.41.2
torch==2.3.1
wikipedia==1.4.0
yfinance==1.2.0



